reorganized zns utils and tests
