+++ src/utils/znsUtils.ts
